Clean up scheduler utils, move to ir_utils.
