+++ torch/csrc/jit/codegen/cuda/scheduler/utils.h
@@ -39,36 +39,6 @@ size_t mergeNonReduction(
     TensorView* tv,
     const std::unordered_set<IterDomain*>& dont_merge = {});
 
-// Makes rfactor generic with reduction ops and Welford
-TensorView* rfactorHelper(TensorView* red_tv, const std::vector<int>& axes);
-
-// Return immediate producers of tv
-std::vector<TensorView*> producerTvsOf(TensorView* tv);
-
-// Return immediate consumers of tv
-std::vector<TensorView*> consumerTvsOf(TensorView* tv);
-
-// Return immediate producers of tvs (can return tvs input)
-std::vector<TensorView*> producerTvsOf(const std::vector<TensorView*>& tvs);
-
-// Return immediate consumers of tvs (can return tvs input)
-std::vector<TensorView*> consumerTvsOf(const std::vector<TensorView*>& tvs);
-
-// Returns producers of tv that are inputs of fusion
-std::vector<TensorView*> inputTvsOf(TensorView* tv);
-
-// Returns consumers of tv that are outputs of fusion
-std::vector<TensorView*> outputTvsOf(TensorView* tv);
-
-// Returns producers of tvs that are inputs of fusion
-std::vector<TensorView*> inputTvsOf(std::vector<TensorView*> tvs);
-
-// Returns consumers of tvs that are outputs of fusion
-std::vector<TensorView*> outputTvsOf(std::vector<TensorView*> tvs);
-
-// returns all tensor views in fusion that are used between outputs and inputs.
-TORCH_CUDA_CU_API std::vector<TensorView*> allTvs(Fusion* fusion);
-
 TORCH_CUDA_CU_API void parallelizeAllLike(
     TensorView* reference_tv,
     const std::vector<TensorView*>& all_tvs);
